+++ server/src/domain/download/download.service.ts
@@ -4,12 +4,12 @@ import { AccessCore, Permission } from 'src/domain/access/access.core';
 import { AssetIdsDto } from 'src/domain/asset/dto/asset-ids.dto';
 import { AuthDto } from 'src/domain/auth/auth.dto';
 import { mimeTypes } from 'src/domain/domain.constant';
-import { CacheControl, HumanReadableSize, ImmichFileResponse, usePagination } from 'src/domain/domain.util';
 import { DownloadArchiveInfo, DownloadInfoDto, DownloadResponseDto } from 'src/domain/download/download.dto';
 import { IAccessRepository } from 'src/domain/repositories/access.repository';
 import { IAssetRepository } from 'src/domain/repositories/asset.repository';
 import { IStorageRepository, ImmichReadStream } from 'src/domain/repositories/storage.repository';
 import { AssetEntity } from 'src/infra/entities/asset.entity';
+import { CacheControl, HumanReadableSize, ImmichFileResponse, usePagination } from 'src/utils';
 
 @Injectable()
 export class DownloadService {
