+++ server/src/domain/shared-link/shared-link.service.ts
@@ -3,7 +3,6 @@ import { AccessCore, Permission } from 'src/domain/access/access.core';
 import { AssetIdsDto } from 'src/domain/asset/dto/asset-ids.dto';
 import { AssetIdErrorReason, AssetIdsResponseDto } from 'src/domain/asset/response-dto/asset-ids-response.dto';
 import { AuthDto } from 'src/domain/auth/auth.dto';
-import { OpenGraphTags } from 'src/domain/domain.util';
 import { IAccessRepository } from 'src/domain/repositories/access.repository';
 import { ICryptoRepository } from 'src/domain/repositories/crypto.repository';
 import { ISharedLinkRepository } from 'src/domain/repositories/shared-link.repository';
@@ -15,6 +14,7 @@ import {
 import { SharedLinkCreateDto, SharedLinkEditDto, SharedLinkPasswordDto } from 'src/domain/shared-link/shared-link.dto';
 import { AssetEntity } from 'src/infra/entities/asset.entity';
 import { SharedLinkEntity, SharedLinkType } from 'src/infra/entities/shared-link.entity';
+import { OpenGraphTags } from 'src/utils';
 
 @Injectable()
 export class SharedLinkService {
