+++ server/src/domain/smart-info/smart-info.service.ts
@@ -1,5 +1,4 @@
 import { Inject, Injectable } from '@nestjs/common';
-import { usePagination } from 'src/domain/domain.util';
 import { JOBS_ASSET_PAGINATION_SIZE, JobName, QueueName } from 'src/domain/job/job.constants';
 import { IBaseJob, IEntityJob } from 'src/domain/job/job.interface';
 import { IAssetRepository, WithoutProperty } from 'src/domain/repositories/asset.repository';
@@ -10,6 +9,7 @@ import { ISearchRepository } from 'src/domain/repositories/search.repository';
 import { ISystemConfigRepository } from 'src/domain/repositories/system-config.repository';
 import { SystemConfigCore } from 'src/domain/system-config/system-config.core';
 import { ImmichLogger } from 'src/infra/logger';
+import { usePagination } from 'src/utils';
 
 @Injectable()
 export class SmartInfoService {
