+++ server/src/immich/app.utils.ts
@@ -15,10 +15,10 @@ import path, { isAbsolute } from 'node:path';
 import { promisify } from 'node:util';
 import { IMMICH_ACCESS_COOKIE, IMMICH_API_KEY_HEADER, IMMICH_API_KEY_NAME } from 'src/domain/auth/auth.constant';
 import { serverVersion } from 'src/domain/domain.constant';
-import { CacheControl, ImmichFileResponse, isConnectionAborted } from 'src/domain/domain.util';
 import { ImmichReadStream } from 'src/domain/repositories/storage.repository';
 import { Metadata } from 'src/immich/app.guard';
 import { ImmichLogger } from 'src/infra/logger';
+import { CacheControl, ImmichFileResponse, isConnectionAborted } from 'src/utils';
 
 type SendFile = Parameters<Response['sendFile']>;
 type SendFileOptions = SendFile[1];
