+++ server/src/infra/repositories/activity.repository.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { DummyValue, GenerateSql } from 'src/decorators';
 import { IActivityRepository } from 'src/domain/repositories/activity.repository';
 import { ActivityEntity } from 'src/infra/entities/activity.entity';
-import { DummyValue, GenerateSql } from 'src/infra/infra.util';
 import { Instrumentation } from 'src/infra/instrumentation';
 import { IsNull, Repository } from 'typeorm';
