+++ server/src/infra/repositories/move.repository.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { DummyValue, GenerateSql } from 'src/decorators';
 import { IMoveRepository, MoveCreate } from 'src/domain/repositories/move.repository';
 import { MoveEntity, PathType } from 'src/infra/entities/move.entity';
-import { DummyValue, GenerateSql } from 'src/infra/infra.util';
 import { Instrumentation } from 'src/infra/instrumentation';
 import { Repository } from 'typeorm';
