+++ server/src/infra/repositories/shared-link.repository.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { DummyValue, GenerateSql } from 'src/decorators';
 import { ISharedLinkRepository } from 'src/domain/repositories/shared-link.repository';
 import { SharedLinkEntity } from 'src/infra/entities/shared-link.entity';
-import { DummyValue, GenerateSql } from 'src/infra/infra.util';
 import { Instrumentation } from 'src/infra/instrumentation';
 import { Repository } from 'typeorm';
