+++ server/src/infra/repositories/user.repository.ts
@@ -1,5 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { DummyValue, GenerateSql } from 'src/decorators';
 import {
   IUserRepository,
   UserFindOptions,
@@ -8,7 +9,6 @@ import {
 } from 'src/domain/repositories/user.repository';
 import { AssetEntity } from 'src/infra/entities/asset.entity';
 import { UserEntity } from 'src/infra/entities/user.entity';
-import { DummyValue, GenerateSql } from 'src/infra/infra.util';
 import { Instrumentation } from 'src/infra/instrumentation';
 import { IsNull, Not, Repository } from 'typeorm';
