+++ server/src/infra/sql-generator/index.ts
@@ -5,10 +5,10 @@ import { Test } from '@nestjs/testing';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { mkdir, rm, writeFile } from 'node:fs/promises';
 import { join } from 'node:path';
+import { GENERATE_SQL_KEY, GenerateSqlQueries } from 'src/decorators';
 import { ISystemConfigRepository } from 'src/domain/repositories/system-config.repository';
 import { databaseConfig } from 'src/infra/database.config';
 import { databaseEntities } from 'src/infra/entities';
-import { GENERATE_SQL_KEY, GenerateSqlQueries } from 'src/infra/infra.util';
 import { AccessRepository } from 'src/infra/repositories/access.repository';
 import { AlbumRepository } from 'src/infra/repositories/album.repository';
 import { ApiKeyRepository } from 'src/infra/repositories/api-key.repository';
